+++ pisces/_modidx.py
@@ -74,7 +74,9 @@ d = { 'settings': { 'branch': 'main',
                                            'pisces.mads_olsen_support.load_saved_keras': ( 'mads_olsen_support.html#load_saved_keras',
                                                                                            'pisces/mads_olsen_support.py'),
                                            'pisces.mads_olsen_support.median': ( 'mads_olsen_support.html#median',
-                                                                                 'pisces/mads_olsen_support.py')},
+                                                                                 'pisces/mads_olsen_support.py'),
+                                           'pisces.mads_olsen_support.pisces_setup': ( 'mads_olsen_support.html#pisces_setup',
+                                                                                       'pisces/mads_olsen_support.py')},
             'pisces.models': { 'pisces.models.LeaveOneOutSplitter': ('models.html#leaveoneoutsplitter', 'pisces/models.py'),
                                'pisces.models.LeaveOneOutSplitter.split': ('models.html#leaveoneoutsplitter.split', 'pisces/models.py'),
                                'pisces.models.MOResUNetPretrained': ('models.html#moresunetpretrained', 'pisces/models.py'),
